+++ src/core-providers/help-message-service/help-message-service.ts
@@ -2,13 +2,13 @@
 
 import { HelpMessageTypes } from './help-message-types';
 
-import { COMPACT_HELP_MESSAGE } from '../../commands/compact/compact-help-message';
-import { DESERIALIZE_HELP_MESSAGE } from '../../commands/deserialize/deserialize-help-message';
-import { EXPAND_HELP_MESSAGE } from '../../commands/expand/expand-help-message';
-import { FLATTEN_HELP_MESSAGE } from '../../commands/flatten/flatten-help-message';
-import { FRAME_HELP_MESSAGE } from '../../commands/frame/frame-help-message';
-import { NORMALIZE_HELP_MESSAGE } from '../../commands/normalize/normalize-help-message';
-import { SERIALIZE_HELP_MESSAGE } from '../../commands/serialize/serialize-help-message';
+import { COMPACT_HELP_MESSAGE } from '../../operations/compact/compact-help-message';
+import { DESERIALIZE_HELP_MESSAGE } from '../../operations/deserialize/deserialize-help-message';
+import { EXPAND_HELP_MESSAGE } from '../../operations/expand/expand-help-message';
+import { FLATTEN_HELP_MESSAGE } from '../../operations/flatten/flatten-help-message';
+import { FRAME_HELP_MESSAGE } from '../../operations/frame/frame-help-message';
+import { NORMALIZE_HELP_MESSAGE } from '../../operations/normalize/normalize-help-message';
+import { SERIALIZE_HELP_MESSAGE } from '../../operations/serialize/serialize-help-message';
 
 const MESSAGE_MAP: Map<HelpMessageTypes, string> = new Map([
     [HelpMessageTypes.COMPACT, COMPACT_HELP_MESSAGE],