+++ src/core-providers/operation-execution-service/option-types.ts
@@ -1,0 +1,7 @@
+// Option types
+
+import { CompactOptions, ExpandOptions, FlattenOptions, FromRdfOptions, NormalizeOptions,
+         ToRdfOptions, FrameOptions } from 'jsonld';
+
+export type OperationOptions = CompactOptions | ExpandOptions | FlattenOptions |
+                               FrameOptions | FromRdfOptions | NormalizeOptions | ToRdfOptions;