+++ src/index.ts
@@ -1,3 +1,10 @@
+#! /usr/bin/env node
 // entrypoint
 
-export * from './main';
+import { container } from './inversify//inversify.config';
+import { TYPES } from './inversify/types';
+import { NodeJSONLD } from './main';
+
+const nodeJsonLd = container.get<NodeJSONLD>(TYPES.NodeJSONLD);
+
+export { nodeJsonLd };