+++ src/models/settings-model.ts
@@ -1,18 +1,18 @@
-// model that provides a sketch of package.json fields
+// // model that provides a sketch of package.json fields
 
-interface Settings {
-    version: string;
-    license: string;
-    repository: string;
-}
+// interface Settings {
+//     version: string;
+//     license: string;
+//     repository: string;
+// }
 
-function emptySettings(): Settings {
-    const retval: Settings = {
-        version: '',
-        license: '',
-        repository: ''
-    };
-    return retval;
-}
+// function emptySettings(): Settings {
+//     const retval: Settings = {
+//         version: '',
+//         license: '',
+//         repository: ''
+//     };
+//     return retval;
+// }
 
-export { Settings, emptySettings };
+// export { Settings, emptySettings };