+++ src/operations/expand/expand-help-message.ts
@@ -1,0 +1,5 @@
+// expand help message
+export const EXPAND_HELP_MESSAGE = 'Expand the JSON or JSON-LD in <sourceFile>, and write the result to [targetFile]. ' +
+                                   'If not specified, the name of the target file is targetFile.json. ' +
+                                   'For more information: http://json-ld.org/spec/latest/json-ld/#expanded-document-form';
+